requirements.txt
numpy
torch
whisper
pydub
beautifulsoup4
pdfkit
Flask
python-dotenv
tqdm
yt-dlp
instaloader